requirements.txt
rapidfuzz~=3.12.2
pytesseract~=0.3.13
numpy~=1.26.4
PyMuPDF~=1.25.3
Pillow~=10.4.0
pyzbar~=0.1.9
opencv-python~=4.11.0.86
web-browser
tkinter
customtkinter~=5.2.2
DateTime~=5.5
python-dateutil~=2.9.0.post0
logging~=0.4.9.6